Sync Plans in the LocationTaskIntentService and make it main Service
