+++ opensrp-reveal/src/main/java/org/smartregister/reveal/interactor/LoginInteractor.java
@@ -1,10 +1,10 @@
 package org.smartregister.reveal.interactor;
 
-import org.smartregister.job.PlanIntentServiceJob;
 import org.smartregister.job.PullUniqueIdsServiceJob;
 import org.smartregister.job.SyncServiceJob;
 import org.smartregister.login.interactor.BaseLoginInteractor;
 import org.smartregister.reveal.BuildConfig;
+import org.smartregister.reveal.job.LocationTaskServiceJob;
 import org.smartregister.reveal.util.TestDataUtils;
 import org.smartregister.reveal.util.Utils;
 import org.smartregister.view.contract.BaseLoginContract;
@@ -19,7 +19,7 @@ public class LoginInteractor extends BaseLoginInteractor implements BaseLoginCon
 
     @Override
     protected void scheduleJobsPeriodically() {
-        PlanIntentServiceJob.scheduleJob(PlanIntentServiceJob.TAG, TimeUnit.MINUTES.toMillis(
+        LocationTaskServiceJob.scheduleJob(LocationTaskServiceJob.TAG, TimeUnit.MINUTES.toMillis(
                 BuildConfig.SYNC_INTERVAL_IN_MINUTES), getFlexValue(BuildConfig.SYNC_INTERVAL_IN_MINUTES));
 
         PullUniqueIdsServiceJob.scheduleJob(SyncServiceJob.TAG, TimeUnit.MINUTES.toMillis(
