--- opensrp-reveal/src/main/java/org/smartregister/reveal/job/RevealPlanServiceJob.java
@@ -1,20 +1,0 @@
-package org.smartregister.reveal.job;
-
-import android.content.Intent;
-import android.support.annotation.NonNull;
-
-import org.smartregister.AllConstants;
-import org.smartregister.job.BaseJob;
-import org.smartregister.job.PlanIntentServiceJob;
-import org.smartregister.reveal.sync.RevealPlanIntentService;
-
-public class RevealPlanServiceJob extends PlanIntentServiceJob {
-
-    @NonNull
-    @Override
-    protected Result onRunJob(@NonNull Params params) {
-        Intent intent = new Intent(getApplicationContext(), RevealPlanIntentService.class);
-        getApplicationContext().startService(intent);
-        return params != null && params.getExtras().getBoolean(AllConstants.INTENT_KEY.TO_RESCHEDULE, false) ? Result.RESCHEDULE : Result.SUCCESS;
-    }
-}
