+++ opensrp-reveal/src/main/java/org/smartregister/reveal/presenter/TaskRegisterFragmentPresenter.java
@@ -121,7 +121,7 @@ public class TaskRegisterFragmentPresenter extends BaseFormFragmentPresenter imp
      */
     private Pair<String, String[]> getMainCondition() {
         Location operationalArea = Utils.getOperationalAreaLocation(prefsUtil.getCurrentOperationalArea());
-        String whereClause = String.format("%s = ? AND %s = ?", DatabaseKeys.GROUPID, DatabaseKeys.CAMPAIGN_ID);
+        String whereClause = String.format("%s = ? AND %s = ?", DatabaseKeys.GROUPID, DatabaseKeys.PLAN_ID);
         return new Pair<>(whereClause, new String[]{operationalArea == null ?
                 null : operationalArea.getId(), prefsUtil.getCurrentPlanId()});
     }
